requirements-core.txt
# ===============================
#  ЛЕГКІ БІБЛІОТЕКИ (ОСНОВА)
#  Для швидкого встановлення
# ===============================

# --- Для інтерфейсу ---
PySide6==6.9.2
PySide6-Addons==6.9.2
PySide6-Essentials==6.9.2

# --- Для роботи з мережею та системою ---
requests==2.32.3
python-dotenv
psutil==7.0.0

# --- Для Google API (автентифікація, ліцензії) ---
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1